add cast
